React Routing Completed, about to start Firebase
