+++ package.json
@@ -6,8 +6,10 @@
     "@testing-library/jest-dom": "5.16.1",
     "@testing-library/react": "12.1.2",
     "@testing-library/user-event": "13.5.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "firebase": "^9.22.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.12.0",
     "react-scripts": "5.0.0",
     "sass": "^1.62.1",
     "web-vitals": "2.1.4"
