+++ src/App.js
@@ -1,41 +1,23 @@
-import Directory from "./components/directory/directory.components";
+import { Routes, Route } from "react-router-dom";
+import Home from "./routes/Home/home.component";
+import Navigation from "./routes/navigation/navigation.components";
 
-
+const Shop = () =>{
+  return (
+    <>
+    <div><h1>I am Shop</h1></div>
+    </>
+  )
+}
 
 const App = () => {
-  const categories = [
-    {
-      id: 1,
-      title: "hats",
-      imageUrl: "https://i.ibb.co/cvpntL1/hats.png",
-    },
-    {
-      id: 2,
-      title: "jackets",
-      imageUrl: "https://i.ibb.co/px2tCc3/jackets.png",
-    },
-    {
-      id: 3,
-      title: "sneakers",
-      imageUrl: "https://i.ibb.co/0jqHpnp/sneakers.png",
-    },
-    {
-      id: 4,
-      title: "womens",
-      imageUrl: "https://i.ibb.co/GCCdy8t/womens.png",
-    },
-    {
-      id: 5,
-      title: "mens",
-      imageUrl: "https://i.ibb.co/R70vBrQ/men.png",
-    },
-  ];
-
   return (
-
-    <>
-    <Directory categories = {categories} />
-    </>
+    <Routes>
+      <Route path="/" element={<Navigation />}>
+        <Route index element={<Home />} />
+        <Route path = 'shop' element = {<Shop/>}/>
+      </Route>
+    </Routes>
   );
 };
