+++ src/components/category-item/category-item.components.jsx
@@ -1,7 +1,7 @@
 import './category-item.styles.scss'
 
 const CategoryItem = ({category}) => {
-    const {title , imageUrl} = category
+    const {title , imageUrl , id} = category
   return (
     <div  className="category-container">
           <div className="background-image" style={{backgroundImage : `url(${imageUrl})`}}/>
