+++ src/routes/navigation/navigation.components.jsx
@@ -1,0 +1,22 @@
+import { Outlet, Link } from "react-router-dom";
+import {ReactComponent as CrwnLogo } from '../../assets/crown(1).svg'
+import './navigation.styles.scss'
+const Navigation = () => {
+  return (
+    <>
+      <div className="navigation">
+        <Link className="logo-container" to="/">
+          <CrwnLogo className = 'logo'/>
+        </Link>
+        <div className="nav-links-container">
+          <Link className="nav-link" to="/shop">
+            SHOP
+          </Link>
+        </div>
+      </div>
+      <Outlet />
+    </>
+  );
+};
+
+export default Navigation;
